Hosted Activity rework
We are now destroying the hosted activity within the
LocalActivityManager when the MapFragment's view
gets unloaded (onViewDestroy).

Since our MapFragment's view hosts a MapView which
which is controlled by the LocalActivityManager it only
makes sense to call a destroyActivity on it when the
onDestroyView is called.  When our MapFragment's
view gets recreated then the createActivity gets called
again and everything works nicely together.